% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_os}
\alias{get_os}
\title{Get the OS from R}
\usage{
get_os()
}
\description{
Get the type of the operating system. The returned value is used in configuring
parallel computation for the implemented algorithms.
}
\references{
This function is adapted from \url{https://www.r-bloggers.com/identifying-the-os-from-r/}
}
